requirements.txt
matplotlib
numpy
opencv-python
openslide-python
packaging
Pillow
tqdm